+++ action_scripts/close.gd
@@ -14,4 +14,6 @@ func _run_action() -> void:
 	Global.set_file_path("")
 	Global.set_editor_text("")
 	Global.set_editor_disabled(true)
+	Global.get_editor().clear_bookmarked_lines()
+	Global.get_editor().type_timer_timeout.emit()
 	Signals.check_options.emit()